Switch to using gliderlabs/alpine
Using scratch as a baseimage is bad because of the PID 1 zombie reaping problem. See https://blog.phusion.nl/2015/01/20/docker-and-the-pid-1-zombie-reaping-problem for more details on it.
gliderlabs/alpine is a great solution for this, as it is based on Alpine Linux, is only 5 megabytes, and has a full featured init process.